Make embedded link method reusable
